requirements.txt
python-telegram-bot==20.6
mysql-connector-python
python-dotenv
